requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-rest-swagger
drf-yasg
